Module aws_sdk_elasticsearch::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptInboundCrossClusterSearchConnection

Fluent builder constructing a request to AcceptInboundCrossClusterSearchConnection.

AddTags

Fluent builder constructing a request to AddTags.

AssociatePackage

Fluent builder constructing a request to AssociatePackage.

CancelElasticsearchServiceSoftwareUpdate

Fluent builder constructing a request to CancelElasticsearchServiceSoftwareUpdate.

CreateElasticsearchDomain

Fluent builder constructing a request to CreateElasticsearchDomain.

CreateOutboundCrossClusterSearchConnection

Fluent builder constructing a request to CreateOutboundCrossClusterSearchConnection.

CreatePackage

Fluent builder constructing a request to CreatePackage.

DeleteElasticsearchDomain

Fluent builder constructing a request to DeleteElasticsearchDomain.

DeleteElasticsearchServiceRole

Fluent builder constructing a request to DeleteElasticsearchServiceRole.

DeleteInboundCrossClusterSearchConnection

Fluent builder constructing a request to DeleteInboundCrossClusterSearchConnection.

DeleteOutboundCrossClusterSearchConnection

Fluent builder constructing a request to DeleteOutboundCrossClusterSearchConnection.

DeletePackage

Fluent builder constructing a request to DeletePackage.

DescribeDomainAutoTunes

Fluent builder constructing a request to DescribeDomainAutoTunes.

DescribeElasticsearchDomain

Fluent builder constructing a request to DescribeElasticsearchDomain.

DescribeElasticsearchDomainConfig

Fluent builder constructing a request to DescribeElasticsearchDomainConfig.

DescribeElasticsearchDomains

Fluent builder constructing a request to DescribeElasticsearchDomains.

DescribeElasticsearchInstanceTypeLimits

Fluent builder constructing a request to DescribeElasticsearchInstanceTypeLimits.

DescribeInboundCrossClusterSearchConnections

Fluent builder constructing a request to DescribeInboundCrossClusterSearchConnections.

DescribeOutboundCrossClusterSearchConnections

Fluent builder constructing a request to DescribeOutboundCrossClusterSearchConnections.

DescribePackages

Fluent builder constructing a request to DescribePackages.

DescribeReservedElasticsearchInstanceOfferings

Fluent builder constructing a request to DescribeReservedElasticsearchInstanceOfferings.

DescribeReservedElasticsearchInstances

Fluent builder constructing a request to DescribeReservedElasticsearchInstances.

DissociatePackage

Fluent builder constructing a request to DissociatePackage.

GetCompatibleElasticsearchVersions

Fluent builder constructing a request to GetCompatibleElasticsearchVersions.

GetPackageVersionHistory

Fluent builder constructing a request to GetPackageVersionHistory.

GetUpgradeHistory

Fluent builder constructing a request to GetUpgradeHistory.

GetUpgradeStatus

Fluent builder constructing a request to GetUpgradeStatus.

ListDomainNames

Fluent builder constructing a request to ListDomainNames.

ListDomainsForPackage

Fluent builder constructing a request to ListDomainsForPackage.

ListElasticsearchInstanceTypes

Fluent builder constructing a request to ListElasticsearchInstanceTypes.

ListElasticsearchVersions

Fluent builder constructing a request to ListElasticsearchVersions.

ListPackagesForDomain

Fluent builder constructing a request to ListPackagesForDomain.

ListTags

Fluent builder constructing a request to ListTags.

PurchaseReservedElasticsearchInstanceOffering

Fluent builder constructing a request to PurchaseReservedElasticsearchInstanceOffering.

RejectInboundCrossClusterSearchConnection

Fluent builder constructing a request to RejectInboundCrossClusterSearchConnection.

RemoveTags

Fluent builder constructing a request to RemoveTags.

StartElasticsearchServiceSoftwareUpdate

Fluent builder constructing a request to StartElasticsearchServiceSoftwareUpdate.

UpdateElasticsearchDomainConfig

Fluent builder constructing a request to UpdateElasticsearchDomainConfig.

UpdatePackage

Fluent builder constructing a request to UpdatePackage.

UpgradeElasticsearchDomain

Fluent builder constructing a request to UpgradeElasticsearchDomain.