Struct aws_sdk_elasticsearch::client::fluent_builders::UpdateElasticsearchDomainConfig

pub struct UpdateElasticsearchDomainConfig<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateElasticsearchDomainConfig<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateElasticsearchDomainConfigOutput, SdkError<UpdateElasticsearchDomainConfigError>> where
R::Policy: SmithyRetryPolicy<UpdateElasticsearchDomainConfigInputOperationOutputAlias, UpdateElasticsearchDomainConfigOutput, UpdateElasticsearchDomainConfigError, UpdateElasticsearchDomainConfigInputOperationRetryAlias>,

pub fn domain_name(self, inp: impl Into<String>) -> Self

The name of the Elasticsearch domain that you are updating.

pub fn set_domain_name(self, input: Option<String>) -> Self

pub fn elasticsearch_cluster_config(
self,
inp: ElasticsearchClusterConfig
) -> Self

The type and number of instances to instantiate for the domain cluster.

pub fn set_elasticsearch_cluster_config(
self,
input: Option<ElasticsearchClusterConfig>
) -> Self

pub fn ebs_options(self, inp: EbsOptions) -> Self

Specify the type and size of the EBS volume that you want to use.

pub fn set_ebs_options(self, input: Option<EbsOptions>) -> Self

pub fn snapshot_options(self, inp: SnapshotOptions) -> Self

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

pub fn set_snapshot_options(self, input: Option<SnapshotOptions>) -> Self

pub fn vpc_options(self, inp: VpcOptions) -> Self

Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

pub fn set_vpc_options(self, input: Option<VpcOptions>) -> Self

pub fn cognito_options(self, inp: CognitoOptions) -> Self

Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

pub fn set_cognito_options(self, input: Option<CognitoOptions>) -> Self

pub fn advanced_options(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to AdvancedOptions.

To override the contents of this collection use set_advanced_options.

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

pub fn set_advanced_options(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn access_policies(self, inp: impl Into<String>) -> Self

IAM access policy as a JSON-formatted string.

pub fn set_access_policies(self, input: Option<String>) -> Self

pub fn log_publishing_options(
self,
k: impl Into<LogType>,
v: impl Into<LogPublishingOption>
) -> Self

Adds a key-value pair to LogPublishingOptions.

To override the contents of this collection use set_log_publishing_options.

Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

pub fn set_log_publishing_options(
self,
input: Option<HashMap<LogType, LogPublishingOption>>
) -> Self

pub fn domain_endpoint_options(self, inp: DomainEndpointOptions) -> Self

Options to specify configuration that will be applied to the domain endpoint.

pub fn set_domain_endpoint_options(
self,
input: Option<DomainEndpointOptions>
) -> Self

pub fn advanced_security_options(
self,
inp: AdvancedSecurityOptionsInput
) -> Self

Specifies advanced security options.

pub fn set_advanced_security_options(
self,
input: Option<AdvancedSecurityOptionsInput>
) -> Self

pub fn node_to_node_encryption_options(
self,
inp: NodeToNodeEncryptionOptions
) -> Self

Specifies the NodeToNodeEncryptionOptions.

pub fn set_node_to_node_encryption_options(
self,
input: Option<NodeToNodeEncryptionOptions>
) -> Self

pub fn encryption_at_rest_options(self, inp: EncryptionAtRestOptions) -> Self

Specifies the Encryption At Rest Options.

pub fn set_encryption_at_rest_options(
self,
input: Option<EncryptionAtRestOptions>
) -> Self

pub fn auto_tune_options(self, inp: AutoTuneOptions) -> Self

Specifies Auto-Tune options.

pub fn set_auto_tune_options(self, input: Option<AutoTuneOptions>) -> Self