Struct aws_sdk_elasticsearch::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonElasticsearchService2015.

This client allows ergonomic access to a AmazonElasticsearchService2015-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_inbound_cross_cluster_search_connection(
&self
) -> AcceptInboundCrossClusterSearchConnection<C, M, R>

pub fn add_tags(&self) -> AddTags<C, M, R>

pub fn associate_package(&self) -> AssociatePackage<C, M, R>

pub fn cancel_elasticsearch_service_software_update(
&self
) -> CancelElasticsearchServiceSoftwareUpdate<C, M, R>

pub fn create_elasticsearch_domain(&self) -> CreateElasticsearchDomain<C, M, R>

pub fn create_outbound_cross_cluster_search_connection(
&self
) -> CreateOutboundCrossClusterSearchConnection<C, M, R>

pub fn create_package(&self) -> CreatePackage<C, M, R>

pub fn delete_elasticsearch_domain(&self) -> DeleteElasticsearchDomain<C, M, R>

pub fn delete_elasticsearch_service_role(
&self
) -> DeleteElasticsearchServiceRole<C, M, R>

pub fn delete_inbound_cross_cluster_search_connection(
&self
) -> DeleteInboundCrossClusterSearchConnection<C, M, R>

pub fn delete_outbound_cross_cluster_search_connection(
&self
) -> DeleteOutboundCrossClusterSearchConnection<C, M, R>

pub fn delete_package(&self) -> DeletePackage<C, M, R>

pub fn describe_domain_auto_tunes(&self) -> DescribeDomainAutoTunes<C, M, R>

pub fn describe_elasticsearch_domain(
&self
) -> DescribeElasticsearchDomain<C, M, R>

pub fn describe_elasticsearch_domain_config(
&self
) -> DescribeElasticsearchDomainConfig<C, M, R>

pub fn describe_elasticsearch_domains(
&self
) -> DescribeElasticsearchDomains<C, M, R>

pub fn describe_elasticsearch_instance_type_limits(
&self
) -> DescribeElasticsearchInstanceTypeLimits<C, M, R>

pub fn describe_inbound_cross_cluster_search_connections(
&self
) -> DescribeInboundCrossClusterSearchConnections<C, M, R>

pub fn describe_outbound_cross_cluster_search_connections(
&self
) -> DescribeOutboundCrossClusterSearchConnections<C, M, R>

pub fn describe_packages(&self) -> DescribePackages<C, M, R>

pub fn describe_reserved_elasticsearch_instance_offerings(
&self
) -> DescribeReservedElasticsearchInstanceOfferings<C, M, R>

pub fn describe_reserved_elasticsearch_instances(
&self
) -> DescribeReservedElasticsearchInstances<C, M, R>

pub fn dissociate_package(&self) -> DissociatePackage<C, M, R>

pub fn get_compatible_elasticsearch_versions(
&self
) -> GetCompatibleElasticsearchVersions<C, M, R>

pub fn get_package_version_history(&self) -> GetPackageVersionHistory<C, M, R>

pub fn get_upgrade_history(&self) -> GetUpgradeHistory<C, M, R>

pub fn get_upgrade_status(&self) -> GetUpgradeStatus<C, M, R>

pub fn list_domain_names(&self) -> ListDomainNames<C, M, R>

pub fn list_domains_for_package(&self) -> ListDomainsForPackage<C, M, R>

pub fn list_elasticsearch_instance_types(
&self
) -> ListElasticsearchInstanceTypes<C, M, R>

pub fn list_elasticsearch_versions(&self) -> ListElasticsearchVersions<C, M, R>

pub fn list_packages_for_domain(&self) -> ListPackagesForDomain<C, M, R>

pub fn list_tags(&self) -> ListTags<C, M, R>

pub fn purchase_reserved_elasticsearch_instance_offering(
&self
) -> PurchaseReservedElasticsearchInstanceOffering<C, M, R>

pub fn reject_inbound_cross_cluster_search_connection(
&self
) -> RejectInboundCrossClusterSearchConnection<C, M, R>

pub fn remove_tags(&self) -> RemoveTags<C, M, R>

pub fn start_elasticsearch_service_software_update(
&self
) -> StartElasticsearchServiceSoftwareUpdate<C, M, R>

pub fn update_elasticsearch_domain_config(
&self
) -> UpdateElasticsearchDomainConfig<C, M, R>

pub fn update_package(&self) -> UpdatePackage<C, M, R>

pub fn upgrade_elasticsearch_domain(
&self
) -> UpgradeElasticsearchDomain<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self