Module aws_sdk_elasticsearch::error

Modules

access_denied_exception

See AccessDeniedException

base_exception

See BaseException

conflict_exception

See ConflictException

disabled_operation_exception

See DisabledOperationException

internal_exception

See InternalException

invalid_pagination_token_exception

See InvalidPaginationTokenException

invalid_type_exception

See InvalidTypeException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

validation_exception

See ValidationException

Structs

AcceptInboundCrossClusterSearchConnectionError

AccessDeniedException

An error occurred because user does not have permissions to access the resource. Returns HTTP status code 403.

AddTagsError

AssociatePackageError

BaseException

An error occurred while processing the request.

CancelElasticsearchServiceSoftwareUpdateError

ConflictException

An error occurred because the client attempts to remove a resource that is currently in use. Returns HTTP status code 409.

CreateElasticsearchDomainError

CreateOutboundCrossClusterSearchConnectionError

CreatePackageError

DeleteElasticsearchDomainError

DeleteElasticsearchServiceRoleError

DeleteInboundCrossClusterSearchConnectionError

DeleteOutboundCrossClusterSearchConnectionError

DeletePackageError

DescribeDomainAutoTunesError

DescribeElasticsearchDomainConfigError

DescribeElasticsearchDomainError

DescribeElasticsearchDomainsError

DescribeElasticsearchInstanceTypeLimitsError

DescribeInboundCrossClusterSearchConnectionsError

DescribeOutboundCrossClusterSearchConnectionsError

DescribePackagesError

DescribeReservedElasticsearchInstanceOfferingsError

DescribeReservedElasticsearchInstancesError

DisabledOperationException

An error occured because the client wanted to access a not supported operation. Gives http status code of 409.

DissociatePackageError

GetCompatibleElasticsearchVersionsError

GetPackageVersionHistoryError

GetUpgradeHistoryError

GetUpgradeStatusError

InternalException

The request processing has failed because of an unknown error, exception or failure (the failure is internal to the service) . Gives http status code of 500.

InvalidPaginationTokenException

The request processing has failed because of invalid pagination token provided by customer. Returns an HTTP status code of 400.

InvalidTypeException

An exception for trying to create or access sub-resource that is either invalid or not supported. Gives http status code of 409.

LimitExceededException

An exception for trying to create more than allowed resources or sub-resources. Gives http status code of 409.

ListDomainNamesError

ListDomainsForPackageError

ListElasticsearchInstanceTypesError

ListElasticsearchVersionsError

ListPackagesForDomainError

ListTagsError

PurchaseReservedElasticsearchInstanceOfferingError

RejectInboundCrossClusterSearchConnectionError

RemoveTagsError

ResourceAlreadyExistsException

An exception for creating a resource that already exists. Gives http status code of 400.

ResourceNotFoundException

An exception for accessing or deleting a resource that does not exist. Gives http status code of 400.

StartElasticsearchServiceSoftwareUpdateError

UpdateElasticsearchDomainConfigError

UpdatePackageError

UpgradeElasticsearchDomainError

ValidationException

An exception for missing / invalid input fields. Gives http status code of 400.

Enums

AcceptInboundCrossClusterSearchConnectionErrorKind

AddTagsErrorKind

AssociatePackageErrorKind

CancelElasticsearchServiceSoftwareUpdateErrorKind

CreateElasticsearchDomainErrorKind

CreateOutboundCrossClusterSearchConnectionErrorKind

CreatePackageErrorKind

DeleteElasticsearchDomainErrorKind

DeleteElasticsearchServiceRoleErrorKind

DeleteInboundCrossClusterSearchConnectionErrorKind

DeleteOutboundCrossClusterSearchConnectionErrorKind

DeletePackageErrorKind

DescribeDomainAutoTunesErrorKind

DescribeElasticsearchDomainConfigErrorKind

DescribeElasticsearchDomainErrorKind

DescribeElasticsearchDomainsErrorKind

DescribeElasticsearchInstanceTypeLimitsErrorKind

DescribeInboundCrossClusterSearchConnectionsErrorKind

DescribeOutboundCrossClusterSearchConnectionsErrorKind

DescribePackagesErrorKind

DescribeReservedElasticsearchInstanceOfferingsErrorKind

DescribeReservedElasticsearchInstancesErrorKind

DissociatePackageErrorKind

GetCompatibleElasticsearchVersionsErrorKind

GetPackageVersionHistoryErrorKind

GetUpgradeHistoryErrorKind

GetUpgradeStatusErrorKind

ListDomainNamesErrorKind

ListDomainsForPackageErrorKind

ListElasticsearchInstanceTypesErrorKind

ListElasticsearchVersionsErrorKind

ListPackagesForDomainErrorKind

ListTagsErrorKind

PurchaseReservedElasticsearchInstanceOfferingErrorKind

RejectInboundCrossClusterSearchConnectionErrorKind

RemoveTagsErrorKind

StartElasticsearchServiceSoftwareUpdateErrorKind

UpdateElasticsearchDomainConfigErrorKind

UpdatePackageErrorKind

UpgradeElasticsearchDomainErrorKind