Module aws_sdk_elasticsearch::input

Modules

accept_inbound_cross_cluster_search_connection_input

See AcceptInboundCrossClusterSearchConnectionInput

add_tags_input

See AddTagsInput

associate_package_input

See AssociatePackageInput

cancel_elasticsearch_service_software_update_input

See CancelElasticsearchServiceSoftwareUpdateInput

create_elasticsearch_domain_input

See CreateElasticsearchDomainInput

create_outbound_cross_cluster_search_connection_input

See CreateOutboundCrossClusterSearchConnectionInput

create_package_input

See CreatePackageInput

delete_elasticsearch_domain_input

See DeleteElasticsearchDomainInput

delete_elasticsearch_service_role_input

See DeleteElasticsearchServiceRoleInput

delete_inbound_cross_cluster_search_connection_input

See DeleteInboundCrossClusterSearchConnectionInput

delete_outbound_cross_cluster_search_connection_input

See DeleteOutboundCrossClusterSearchConnectionInput

delete_package_input

See DeletePackageInput

describe_domain_auto_tunes_input

See DescribeDomainAutoTunesInput

describe_elasticsearch_domain_config_input

See DescribeElasticsearchDomainConfigInput

describe_elasticsearch_domain_input

See DescribeElasticsearchDomainInput

describe_elasticsearch_domains_input

See DescribeElasticsearchDomainsInput

describe_elasticsearch_instance_type_limits_input

See DescribeElasticsearchInstanceTypeLimitsInput

describe_inbound_cross_cluster_search_connections_input

See DescribeInboundCrossClusterSearchConnectionsInput

describe_outbound_cross_cluster_search_connections_input

See DescribeOutboundCrossClusterSearchConnectionsInput

describe_packages_input

See DescribePackagesInput

describe_reserved_elasticsearch_instance_offerings_input

See DescribeReservedElasticsearchInstanceOfferingsInput

describe_reserved_elasticsearch_instances_input

See DescribeReservedElasticsearchInstancesInput

dissociate_package_input

See DissociatePackageInput

get_compatible_elasticsearch_versions_input

See GetCompatibleElasticsearchVersionsInput

get_package_version_history_input

See GetPackageVersionHistoryInput

get_upgrade_history_input

See GetUpgradeHistoryInput

get_upgrade_status_input

See GetUpgradeStatusInput

list_domain_names_input

See ListDomainNamesInput

list_domains_for_package_input

See ListDomainsForPackageInput

list_elasticsearch_instance_types_input

See ListElasticsearchInstanceTypesInput

list_elasticsearch_versions_input

See ListElasticsearchVersionsInput

list_packages_for_domain_input

See ListPackagesForDomainInput

list_tags_input

See ListTagsInput

purchase_reserved_elasticsearch_instance_offering_input

See PurchaseReservedElasticsearchInstanceOfferingInput

reject_inbound_cross_cluster_search_connection_input

See RejectInboundCrossClusterSearchConnectionInput

remove_tags_input

See RemoveTagsInput

start_elasticsearch_service_software_update_input

See StartElasticsearchServiceSoftwareUpdateInput

update_elasticsearch_domain_config_input

See UpdateElasticsearchDomainConfigInput

update_package_input

See UpdatePackageInput

upgrade_elasticsearch_domain_input

See UpgradeElasticsearchDomainInput

Structs

AcceptInboundCrossClusterSearchConnectionInput

Container for the parameters to the AcceptInboundCrossClusterSearchConnection operation.

AddTagsInput

Container for the parameters to the AddTags operation. Specify the tags that you want to attach to the Elasticsearch domain.

AssociatePackageInput

CancelElasticsearchServiceSoftwareUpdateInput

Container for the parameters to the CancelElasticsearchServiceSoftwareUpdate operation. Specifies the name of the Elasticsearch domain that you wish to cancel a service software update on.

CreateElasticsearchDomainInput

CreateOutboundCrossClusterSearchConnectionInput

Container for the parameters to the CreateOutboundCrossClusterSearchConnection operation.

CreatePackageInput

DeleteElasticsearchDomainInput

Container for the parameters to the DeleteElasticsearchDomain operation. Specifies the name of the Elasticsearch domain that you want to delete.

DeleteElasticsearchServiceRoleInput

DeleteInboundCrossClusterSearchConnectionInput

Container for the parameters to the DeleteInboundCrossClusterSearchConnection operation.

DeleteOutboundCrossClusterSearchConnectionInput

Container for the parameters to the DeleteOutboundCrossClusterSearchConnection operation.

DeletePackageInput

DescribeDomainAutoTunesInput

Container for the parameters to the DescribeDomainAutoTunes operation.

DescribeElasticsearchDomainConfigInput

Container for the parameters to the DescribeElasticsearchDomainConfig operation. Specifies the domain name for which you want configuration information.

DescribeElasticsearchDomainInput

Container for the parameters to the DescribeElasticsearchDomain operation.

DescribeElasticsearchDomainsInput

Container for the parameters to the DescribeElasticsearchDomains operation. By default, the API returns the status of all Elasticsearch domains.

DescribeElasticsearchInstanceTypeLimitsInput

DescribeInboundCrossClusterSearchConnectionsInput

Container for the parameters to the DescribeInboundCrossClusterSearchConnections operation.

DescribeOutboundCrossClusterSearchConnectionsInput

Container for the parameters to the DescribeOutboundCrossClusterSearchConnections operation.

DescribePackagesInput

DescribeReservedElasticsearchInstanceOfferingsInput

Container for parameters to DescribeReservedElasticsearchInstanceOfferings

DescribeReservedElasticsearchInstancesInput

Container for parameters to DescribeReservedElasticsearchInstances

DissociatePackageInput

GetCompatibleElasticsearchVersionsInput

GetPackageVersionHistoryInput

GetUpgradeHistoryInput

GetUpgradeStatusInput

ListDomainNamesInput

ListDomainsForPackageInput

ListElasticsearchInstanceTypesInput

ListElasticsearchVersionsInput

ListPackagesForDomainInput

ListTagsInput

Container for the parameters to the ListTags operation. Specify the ARN for the Elasticsearch domain to which the tags are attached that you want to view are attached.

PurchaseReservedElasticsearchInstanceOfferingInput

Container for parameters to PurchaseReservedElasticsearchInstanceOffering

RejectInboundCrossClusterSearchConnectionInput

Container for the parameters to the RejectInboundCrossClusterSearchConnection operation.

RemoveTagsInput

Container for the parameters to the RemoveTags operation. Specify the ARN for the Elasticsearch domain from which you want to remove the specified TagKey.

StartElasticsearchServiceSoftwareUpdateInput

Container for the parameters to the StartElasticsearchServiceSoftwareUpdate operation. Specifies the name of the Elasticsearch domain that you wish to schedule a service software update on.

UpdateElasticsearchDomainConfigInput

Container for the parameters to the UpdateElasticsearchDomain operation. Specifies the type and number of instances in the domain cluster.

UpdatePackageInput

UpgradeElasticsearchDomainInput