Module aws_sdk_elasticsearch::model

Modules

access_policies_status

See AccessPoliciesStatus

additional_limit

See AdditionalLimit

advanced_options_status

See AdvancedOptionsStatus

advanced_security_options

See AdvancedSecurityOptions

advanced_security_options_input

See AdvancedSecurityOptionsInput

advanced_security_options_status

See AdvancedSecurityOptionsStatus

auto_tune

See AutoTune

auto_tune_details

See AutoTuneDetails

auto_tune_maintenance_schedule

See AutoTuneMaintenanceSchedule

auto_tune_options

See AutoTuneOptions

auto_tune_options_input

See AutoTuneOptionsInput

auto_tune_options_output

See AutoTuneOptionsOutput

auto_tune_options_status

See AutoTuneOptionsStatus

auto_tune_status

See AutoTuneStatus

cognito_options

See CognitoOptions

cognito_options_status

See CognitoOptionsStatus

cold_storage_options

See ColdStorageOptions

compatible_versions_map

See CompatibleVersionsMap

describe_packages_filter

See DescribePackagesFilter

domain_endpoint_options

See DomainEndpointOptions

domain_endpoint_options_status

See DomainEndpointOptionsStatus

domain_info

See DomainInfo

domain_information

See DomainInformation

domain_package_details

See DomainPackageDetails

duration

See Duration

ebs_options

See EbsOptions

ebs_options_status

See EbsOptionsStatus

elasticsearch_cluster_config

See ElasticsearchClusterConfig

elasticsearch_cluster_config_status

See ElasticsearchClusterConfigStatus

elasticsearch_domain_config

See ElasticsearchDomainConfig

elasticsearch_domain_status

See ElasticsearchDomainStatus

elasticsearch_version_status

See ElasticsearchVersionStatus

encryption_at_rest_options

See EncryptionAtRestOptions

encryption_at_rest_options_status

See EncryptionAtRestOptionsStatus

error_details

See ErrorDetails

filter

See Filter

inbound_cross_cluster_search_connection

See InboundCrossClusterSearchConnection

inbound_cross_cluster_search_connection_status

See InboundCrossClusterSearchConnectionStatus

instance_count_limits

See InstanceCountLimits

instance_limits

See InstanceLimits

limits

See Limits

log_publishing_option

See LogPublishingOption

log_publishing_options_status

See LogPublishingOptionsStatus

master_user_options

See MasterUserOptions

node_to_node_encryption_options

See NodeToNodeEncryptionOptions

node_to_node_encryption_options_status

See NodeToNodeEncryptionOptionsStatus

option_status

See OptionStatus

outbound_cross_cluster_search_connection

See OutboundCrossClusterSearchConnection

outbound_cross_cluster_search_connection_status

See OutboundCrossClusterSearchConnectionStatus

package_details

See PackageDetails

package_source

See PackageSource

package_version_history

See PackageVersionHistory

recurring_charge

See RecurringCharge

reserved_elasticsearch_instance

See ReservedElasticsearchInstance

reserved_elasticsearch_instance_offering

See ReservedElasticsearchInstanceOffering

saml_idp

See SamlIdp

saml_options_input

See SamlOptionsInput

saml_options_output

See SamlOptionsOutput

scheduled_auto_tune_details

See ScheduledAutoTuneDetails

service_software_options

See ServiceSoftwareOptions

snapshot_options

See SnapshotOptions

snapshot_options_status

See SnapshotOptionsStatus

storage_type

See StorageType

storage_type_limit

See StorageTypeLimit

tag

See Tag

upgrade_history

See UpgradeHistory

upgrade_step_item

See UpgradeStepItem

vpc_derived_info

See VpcDerivedInfo

vpc_derived_info_status

See VpcDerivedInfoStatus

vpc_options

See VpcOptions

zone_awareness_config

See ZoneAwarenessConfig

Structs

AccessPoliciesStatus

The configured access rules for the domain's document and search endpoints, and the current status of those rules.

AdditionalLimit

AdvancedOptionsStatus

Status of the advanced options for the specified Elasticsearch domain. Currently, the following advanced options are available:

AdvancedSecurityOptions

Specifies the advanced security configuration: whether advanced security is enabled, whether the internal database option is enabled.

AdvancedSecurityOptionsInput

Specifies the advanced security configuration: whether advanced security is enabled, whether the internal database option is enabled, master username and password (if internal database is enabled), and master user ARN (if IAM is enabled).

AdvancedSecurityOptionsStatus

Specifies the status of advanced security options for the specified Elasticsearch domain.

AutoTune

Specifies Auto-Tune type and Auto-Tune action details.

AutoTuneDetails

Specifies details of the Auto-Tune action. See the Developer Guide for more information.

AutoTuneMaintenanceSchedule

Specifies Auto-Tune maitenance schedule. See the Developer Guide for more information.

AutoTuneOptions

Specifies the Auto-Tune options: the Auto-Tune desired state for the domain, rollback state when disabling Auto-Tune options and list of maintenance schedules.

AutoTuneOptionsInput

Specifies the Auto-Tune options: the Auto-Tune desired state for the domain and list of maintenance schedules.

AutoTuneOptionsOutput

Specifies the Auto-Tune options: the Auto-Tune desired state for the domain and list of maintenance schedules.

AutoTuneOptionsStatus

Specifies the status of Auto-Tune options for the specified Elasticsearch domain.

AutoTuneStatus

Provides the current status of the Auto-Tune options.

CognitoOptions

Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

CognitoOptionsStatus

Status of the Cognito options for the specified Elasticsearch domain.

ColdStorageOptions

Specifies settings for cold storage.

CompatibleVersionsMap

DescribePackagesFilter

Filter to apply in DescribePackage response.

DomainEndpointOptions

Options to configure endpoint for the Elasticsearch domain.

DomainEndpointOptionsStatus

The configured endpoint options for the domain and their current status.

DomainInfo

DomainInformation

DomainPackageDetails

Information on a package that is associated with a domain.

Duration

Specifies maintenance schedule duration: duration value and duration unit. See the Developer Guide for more information.

EbsOptions

Options to enable, disable, and specify the properties of EBS storage volumes. For more information, see Configuring EBS-based Storage.

EbsOptionsStatus

Status of the EBS options for the specified Elasticsearch domain.

ElasticsearchClusterConfig

Specifies the configuration for the domain cluster, such as the type and number of instances.

ElasticsearchClusterConfigStatus

Specifies the configuration status for the specified Elasticsearch domain.

ElasticsearchDomainConfig

The configuration of an Elasticsearch domain.

ElasticsearchDomainStatus

The current status of an Elasticsearch domain.

ElasticsearchVersionStatus

Status of the Elasticsearch version options for the specified Elasticsearch domain.

EncryptionAtRestOptions

Specifies the Encryption At Rest Options.

EncryptionAtRestOptionsStatus

Status of the Encryption At Rest options for the specified Elasticsearch domain.

ErrorDetails

Filter

InboundCrossClusterSearchConnection

Specifies details of an inbound connection.

InboundCrossClusterSearchConnectionStatus

Specifies the coonection status of an inbound cross-cluster search connection.

InstanceCountLimits

InstanceLimits

InstanceLimits represents the list of instance related attributes that are available for given InstanceType.

Limits

LogPublishingOption

Log Publishing option that is set for given domain.

LogPublishingOptionsStatus

The configured log publishing options for the domain and their current status.

MasterUserOptions

Credentials for the master user: username and password, ARN, or both.

NodeToNodeEncryptionOptions

Specifies the node-to-node encryption options.

NodeToNodeEncryptionOptionsStatus

Status of the node-to-node encryption options for the specified Elasticsearch domain.

OptionStatus

Provides the current status of the entity.

OutboundCrossClusterSearchConnection

Specifies details of an outbound connection.

OutboundCrossClusterSearchConnectionStatus

Specifies the connection status of an outbound cross-cluster search connection.

PackageDetails

Basic information about a package.

PackageSource

The S3 location for importing the package specified as S3BucketName and S3Key

PackageVersionHistory

Details of a package version.

RecurringCharge

Contains the specific price and frequency of a recurring charges for a reserved Elasticsearch instance, or for a reserved Elasticsearch instance offering.

ReservedElasticsearchInstance

Details of a reserved Elasticsearch instance.

ReservedElasticsearchInstanceOffering

Details of a reserved Elasticsearch instance offering.

SamlIdp

Specifies the SAML Identity Provider's information.

SamlOptionsInput

Specifies the SAML application configuration for the domain.

SamlOptionsOutput

Describes the SAML application configured for the domain.

ScheduledAutoTuneDetails

Specifies details of the scheduled Auto-Tune action. See the Developer Guide for more information.

ServiceSoftwareOptions

The current options of an Elasticsearch domain service software options.

SnapshotOptions

Specifies the time, in UTC format, when the service takes a daily automated snapshot of the specified Elasticsearch domain. Default value is 0 hours.

SnapshotOptionsStatus

Status of a daily automated snapshot.

StorageType

StorageTypes represents the list of storage related types and their attributes

StorageTypeLimit

Limits that are applicable for given storage type.

Tag

Specifies a key value pair for a resource tag.

UpgradeHistory

History of the last 10 Upgrades and Upgrade Eligibility Checks.

UpgradeStepItem

Represents a single step of the Upgrade or Upgrade Eligibility Check workflow.

VpcDerivedInfo

Options to specify the subnets and security groups for VPC endpoint. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

VpcDerivedInfoStatus

Status of the VPC options for the specified Elasticsearch domain.

VpcOptions

Options to specify the subnets and security groups for VPC endpoint. For more information, see VPC Endpoints for Amazon Elasticsearch Service Domains.

ZoneAwarenessConfig

Specifies the zone awareness configuration for the domain cluster, such as the number of availability zones.

Enums

AutoTuneDesiredState

Specifies the Auto-Tune desired state. Valid values are ENABLED, DISABLED.

AutoTuneState

Specifies the Auto-Tune state for the Elasticsearch domain. For valid states see the Developer Guide.

AutoTuneType

Specifies Auto-Tune type. Valid value is SCHEDULED_ACTION.

DeploymentStatus

DescribePackagesFilterName

DomainPackageStatus

EsPartitionInstanceType

EsWarmPartitionInstanceType

InboundCrossClusterSearchConnectionStatusCode

LogType

Type of Log File, it can be one of the following:

OptionState

The state of a requested change. One of the following:

OutboundCrossClusterSearchConnectionStatusCode

PackageStatus

PackageType

ReservedElasticsearchInstancePaymentOption

RollbackOnDisable

Specifies the rollback state while disabling Auto-Tune for the domain. Valid values are NO_ROLLBACK, DEFAULT_ROLLBACK.

ScheduledAutoTuneActionType

Specifies Auto-Tune action type. Valid values are JVM_HEAP_SIZE_TUNING and JVM_YOUNG_GEN_TUNING.

ScheduledAutoTuneSeverityType

Specifies Auto-Tune action severity. Valid values are LOW, MEDIUM and HIGH.

TimeUnit

Specifies the unit of a maintenance schedule duration. Valid value is HOUR. See the Developer Guide for more information.

TlsSecurityPolicy

UpgradeStatus

UpgradeStep

VolumeType

The type of EBS volume, standard, gp2, or io1. See Configuring EBS-based Storagefor more information.