Module aws_sdk_elasticsearch::operation

Structs

AcceptInboundCrossClusterSearchConnection

Allows the destination domain owner to accept an inbound cross-cluster search connection request.

AddTags

Attaches tags to an existing Elasticsearch domain. Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See

AssociatePackage

Associates a package with an Amazon ES domain.

CancelElasticsearchServiceSoftwareUpdate

Cancels a scheduled service software update for an Amazon ES domain. You can only perform this operation before the AutomatedUpdateDate and when the UpdateStatus is in the PENDING_UPDATE state.

CreateElasticsearchDomain

Creates a new Elasticsearch domain. For more information,

CreateOutboundCrossClusterSearchConnection

Creates a new cross-cluster search connection from a source domain to a destination domain.

CreatePackage

Create a package for use with Amazon ES domains.

DeleteElasticsearchDomain

Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.

DeleteElasticsearchServiceRole

Deletes the service-linked role that Elasticsearch Service uses to manage and maintain VPC domains. Role deletion will fail if any existing VPC domains use the role. You must delete any such Elasticsearch domains before deleting the role. See Deleting Elasticsearch Service Role in VPC Endpoints for Amazon Elasticsearch Service Domains.

DeleteInboundCrossClusterSearchConnection

Allows the destination domain owner to delete an existing inbound cross-cluster search connection.

DeleteOutboundCrossClusterSearchConnection

Allows the source domain owner to delete an existing outbound cross-cluster search connection.

DeletePackage

Delete the package.

DescribeDomainAutoTunes

Provides scheduled Auto-Tune action details for the Elasticsearch domain, such as Auto-Tune action type, description, severity, and scheduled date.

DescribeElasticsearchDomain

Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

DescribeElasticsearchDomainConfig

Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.

DescribeElasticsearchDomains

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

DescribeElasticsearchInstanceTypeLimits

DescribeInboundCrossClusterSearchConnections

Lists all the inbound cross-cluster search connections for a destination domain.

DescribeOutboundCrossClusterSearchConnections

Lists all the outbound cross-cluster search connections for a source domain.

DescribePackages

Describes all packages available to Amazon ES. Includes options for filtering, limiting the number of results, and pagination.

DescribeReservedElasticsearchInstanceOfferings

Lists available reserved Elasticsearch instance offerings.

DescribeReservedElasticsearchInstances

Returns information about reserved Elasticsearch instances for this account.

DissociatePackage

Dissociates a package from the Amazon ES domain.

GetCompatibleElasticsearchVersions

GetPackageVersionHistory

Returns a list of versions of the package, along with their creation time and commit message.

GetUpgradeHistory

Retrieves the complete history of the last 10 upgrades that were performed on the domain.

GetUpgradeStatus

Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain.

ListDomainNames

Returns the name of all Elasticsearch domains owned by the current user's account.

ListDomainsForPackage

Lists all Amazon ES domains associated with the package.

ListElasticsearchInstanceTypes

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

ListElasticsearchVersions

List all supported Elasticsearch versions

ListPackagesForDomain

Lists all packages associated with the Amazon ES domain.

ListTags

Returns all tags for the given Elasticsearch domain.

PurchaseReservedElasticsearchInstanceOffering

Allows you to purchase reserved Elasticsearch instances.

RejectInboundCrossClusterSearchConnection

Allows the destination domain owner to reject an inbound cross-cluster search connection request.

RemoveTags

Removes the specified set of tags from the specified Elasticsearch domain.

StartElasticsearchServiceSoftwareUpdate

Schedules a service software update for an Amazon ES domain.

UpdateElasticsearchDomainConfig

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

UpdatePackage

Updates a package for use with Amazon ES domains.

UpgradeElasticsearchDomain

Allows you to either upgrade your domain or perform an Upgrade eligibility check to a compatible Elasticsearch version.