Module aws_sdk_elasticsearch::output

Modules

accept_inbound_cross_cluster_search_connection_output

See AcceptInboundCrossClusterSearchConnectionOutput

add_tags_output

See AddTagsOutput

associate_package_output

See AssociatePackageOutput

cancel_elasticsearch_service_software_update_output

See CancelElasticsearchServiceSoftwareUpdateOutput

create_elasticsearch_domain_output

See CreateElasticsearchDomainOutput

create_outbound_cross_cluster_search_connection_output

See CreateOutboundCrossClusterSearchConnectionOutput

create_package_output

See CreatePackageOutput

delete_elasticsearch_domain_output

See DeleteElasticsearchDomainOutput

delete_elasticsearch_service_role_output

See DeleteElasticsearchServiceRoleOutput

delete_inbound_cross_cluster_search_connection_output

See DeleteInboundCrossClusterSearchConnectionOutput

delete_outbound_cross_cluster_search_connection_output

See DeleteOutboundCrossClusterSearchConnectionOutput

delete_package_output

See DeletePackageOutput

describe_domain_auto_tunes_output

See DescribeDomainAutoTunesOutput

describe_elasticsearch_domain_config_output

See DescribeElasticsearchDomainConfigOutput

describe_elasticsearch_domain_output

See DescribeElasticsearchDomainOutput

describe_elasticsearch_domains_output

See DescribeElasticsearchDomainsOutput

describe_elasticsearch_instance_type_limits_output

See DescribeElasticsearchInstanceTypeLimitsOutput

describe_inbound_cross_cluster_search_connections_output

See DescribeInboundCrossClusterSearchConnectionsOutput

describe_outbound_cross_cluster_search_connections_output

See DescribeOutboundCrossClusterSearchConnectionsOutput

describe_packages_output

See DescribePackagesOutput

describe_reserved_elasticsearch_instance_offerings_output

See DescribeReservedElasticsearchInstanceOfferingsOutput

describe_reserved_elasticsearch_instances_output

See DescribeReservedElasticsearchInstancesOutput

dissociate_package_output

See DissociatePackageOutput

get_compatible_elasticsearch_versions_output

See GetCompatibleElasticsearchVersionsOutput

get_package_version_history_output

See GetPackageVersionHistoryOutput

get_upgrade_history_output

See GetUpgradeHistoryOutput

get_upgrade_status_output

See GetUpgradeStatusOutput

list_domain_names_output

See ListDomainNamesOutput

list_domains_for_package_output

See ListDomainsForPackageOutput

list_elasticsearch_instance_types_output

See ListElasticsearchInstanceTypesOutput

list_elasticsearch_versions_output

See ListElasticsearchVersionsOutput

list_packages_for_domain_output

See ListPackagesForDomainOutput

list_tags_output

See ListTagsOutput

purchase_reserved_elasticsearch_instance_offering_output

See PurchaseReservedElasticsearchInstanceOfferingOutput

reject_inbound_cross_cluster_search_connection_output

See RejectInboundCrossClusterSearchConnectionOutput

remove_tags_output

See RemoveTagsOutput

start_elasticsearch_service_software_update_output

See StartElasticsearchServiceSoftwareUpdateOutput

update_elasticsearch_domain_config_output

See UpdateElasticsearchDomainConfigOutput

update_package_output

See UpdatePackageOutput

upgrade_elasticsearch_domain_output

See UpgradeElasticsearchDomainOutput

Structs

AcceptInboundCrossClusterSearchConnectionOutput

The result of a AcceptInboundCrossClusterSearchConnection operation. Contains details of accepted inbound connection.

AddTagsOutput

AssociatePackageOutput

CancelElasticsearchServiceSoftwareUpdateOutput

The result of a CancelElasticsearchServiceSoftwareUpdate operation. Contains the status of the update.

CreateElasticsearchDomainOutput

The result of a CreateElasticsearchDomain operation. Contains the status of the newly created Elasticsearch domain.

CreateOutboundCrossClusterSearchConnectionOutput

The result of a CreateOutboundCrossClusterSearchConnection request. Contains the details of the newly created cross-cluster search connection.

CreatePackageOutput

DeleteElasticsearchDomainOutput

The result of a DeleteElasticsearchDomain request. Contains the status of the pending deletion, or no status if the domain and all of its resources have been deleted.

DeleteElasticsearchServiceRoleOutput

DeleteInboundCrossClusterSearchConnectionOutput

The result of a DeleteInboundCrossClusterSearchConnection operation. Contains details of deleted inbound connection.

DeleteOutboundCrossClusterSearchConnectionOutput

The result of a DeleteOutboundCrossClusterSearchConnection operation. Contains details of deleted outbound connection.

DeletePackageOutput

DescribeDomainAutoTunesOutput

The result of DescribeDomainAutoTunes request. See the Developer Guide for more information.

DescribeElasticsearchDomainConfigOutput

The result of a DescribeElasticsearchDomainConfig request. Contains the configuration information of the requested domain.

DescribeElasticsearchDomainOutput

The result of a DescribeElasticsearchDomain request. Contains the status of the domain specified in the request.

DescribeElasticsearchDomainsOutput

The result of a DescribeElasticsearchDomains request. Contains the status of the specified domains or all domains owned by the account.

DescribeElasticsearchInstanceTypeLimitsOutput

DescribeInboundCrossClusterSearchConnectionsOutput

The result of a DescribeInboundCrossClusterSearchConnections request. Contains the list of connections matching the filter criteria.

DescribeOutboundCrossClusterSearchConnectionsOutput

The result of a DescribeOutboundCrossClusterSearchConnections request. Contains the list of connections matching the filter criteria.

DescribePackagesOutput

DescribeReservedElasticsearchInstanceOfferingsOutput

Container for results from DescribeReservedElasticsearchInstanceOfferings

DescribeReservedElasticsearchInstancesOutput

Container for results from DescribeReservedElasticsearchInstances

DissociatePackageOutput

GetCompatibleElasticsearchVersionsOutput

GetPackageVersionHistoryOutput

GetUpgradeHistoryOutput

GetUpgradeStatusOutput

ListDomainNamesOutput

The result of a ListDomainNames operation. Contains the names of all Elasticsearch domains owned by this account.

ListDomainsForPackageOutput

ListElasticsearchInstanceTypesOutput

ListElasticsearchVersionsOutput

ListPackagesForDomainOutput

ListTagsOutput

The result of a ListTags operation. Contains tags for all requested Elasticsearch domains.

PurchaseReservedElasticsearchInstanceOfferingOutput

Represents the output of a PurchaseReservedElasticsearchInstanceOffering operation.

RejectInboundCrossClusterSearchConnectionOutput

The result of a RejectInboundCrossClusterSearchConnection operation. Contains details of rejected inbound connection.

RemoveTagsOutput

StartElasticsearchServiceSoftwareUpdateOutput

The result of a StartElasticsearchServiceSoftwareUpdate operation. Contains the status of the update.

UpdateElasticsearchDomainConfigOutput

The result of an UpdateElasticsearchDomain request. Contains the status of the Elasticsearch domain being updated.

UpdatePackageOutput

UpgradeElasticsearchDomainOutput